[Refactor][Refatora o Controller para incluir o ResponseEntity nas rotas e customizar as respostas das requisições para o cliente]

+++ src/main/java/dev/java10x/CadastroDeNinjas/Ninja/NinjaController.java
@@ -1,6 +1,8 @@
 package dev.java10x.CadastroDeNinjas.Ninja;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,31 +25,48 @@ public class NinjaController {
 
     //Adicionar Ninja (CREATE)
     @PostMapping("/criar")
-    public NinjaDTO criarNinja(@RequestBody NinjaDTO ninjaDTO) {
-        return ninjaService.criarNinja(ninjaDTO);
+    public ResponseEntity<String> criarNinja(@RequestBody NinjaDTO ninjaDTO) {
+        NinjaDTO novoNinja = ninjaService.criarNinja(ninjaDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja criado com sucesso:" + novoNinja.getNome() + "ID: " + novoNinja.getId());
     }
 
     //Procurara Ninja por id (READ)
     @GetMapping("/listar/{id}")
-    public NinjaDTO buscarNinjaId(@PathVariable Long id) {
-        return ninjaService.buscarNinjaId(id);
+    public ResponseEntity<?> buscarNinjaId(@PathVariable Long id) {
+
+        NinjaDTO ninjaDTO = ninjaService.buscarNinjaId(id);
+
+        if (ninjaDTO != null) {
+            return ResponseEntity.status(HttpStatus.FOUND).body(ninjaDTO);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja com ID: "+ id + "não encontrado");
     }
 
     //Mostrar todos os Ninjas (READ)
     @GetMapping("/listar")
-    public List<NinjaDTO> listarNinjas() {
-        return ninjaService.listarNinjas();
+    public ResponseEntity<List<NinjaDTO>> listarNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.listarNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     //Alterar dados dos ninjas (UPDATE)
     @PutMapping("/alterar/{id}")
-    public NinjaDTO alterarNinja(@RequestBody NinjaDTO ninjaDTO, @PathVariable Long id) {
-        return ninjaService.atualizarNinja(ninjaDTO, id);
+    public ResponseEntity<?> alterarNinja(@RequestBody NinjaDTO ninjaDTO, @PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.atualizarNinja(ninjaDTO, id);
+        if(ninja != null) {
+            return ResponseEntity.ok(ninja);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja com ID: "+ id + "não encontrado");
     }
 
     //Deletar Ninja (DELETE)
     @DeleteMapping("/deletar/{id}")
-    public void deletarNinjaId(@PathVariable Long id) {
-        ninjaService.deleteNinja(id);
+    public ResponseEntity<String> deletarNinjaId(@PathVariable Long id) {
+        if (ninjaService.buscarNinjaId(id) != null) {
+            ninjaService.deleteNinja(id);
+            return ResponseEntity.ok("Ninja com ID: " + id + "deletado com sucesso");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja com ID: " + id + "não encontrado");
     }
 }
